requirements.txt
autopep8==1.5
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
six==1.14.0
Werkzeug==0.16.1
